point svn:externals to rev 603 of sg3_utils; wscan cleanup

+++ src/ddpt_wscan.c
@@ -280,14 +280,15 @@ query_dev_property(HANDLE hdevice,
         if (verbose > 2) {
             err = GetLastError();
             pr2serr("  IOCTL_STORAGE_QUERY_PROPERTY(Devprop) failed, "
-                    "Error=%ld %s\n", err, get_err_str(err, sizeof(b), b));
+                    "Error=%u %s\n", (unsigned int)err,
+                    get_err_str(err, sizeof(b), b));
         }
         return -ENOSYS;
     }
 
     if (verbose > 3)
-        pr2serr("  IOCTL_STORAGE_QUERY_PROPERTY(DevProp) num_out=%ld\n",
-                num_out);
+        pr2serr("  IOCTL_STORAGE_QUERY_PROPERTY(DevProp) num_out=%u\n",
+                (unsigned int)num_out);
     return 0;
 }
 
@@ -308,11 +309,11 @@ query_dev_uid(HANDLE hdevice,
         if (verbose > 2) {
             err = GetLastError();
             fprintf(stderr, "  IOCTL_STORAGE_QUERY_PROPERTY(DevUid(exists)) "
-                    "failed, Error=%ld %s\n", err,
+                    "failed, Error=%u %s\n", (unsigned int)err,
                     get_err_str(err, sizeof(b), b));
         }
         if (verbose > 3)
-            fprintf(stderr, "      num_out=%ld\n", num_out);
+            fprintf(stderr, "      num_out=%u\n", (unsigned int)num_out);
         /* interpret any error to mean this property doesn't exist */
         return 0;
     }
@@ -324,13 +325,14 @@ query_dev_uid(HANDLE hdevice,
         if (verbose > 2) {
             err = GetLastError();
             fprintf(stderr, "  IOCTL_STORAGE_QUERY_PROPERTY(DevUid) failed, "
-                    "Error=%ld %s\n", err, get_err_str(err, sizeof(b), b));
+                    "Error=%u %s\n", (unsigned int)err,
+                    get_err_str(err, sizeof(b), b));
         }
         return -ENOSYS;
     }
     if (verbose > 3)
-        fprintf(stderr, "  IOCTL_STORAGE_QUERY_PROPERTY(DevUid) num_out=%ld\n",
-                num_out);
+        fprintf(stderr, "  IOCTL_STORAGE_QUERY_PROPERTY(DevUid) num_out=%u\n",
+                (unsigned int)num_out);
     return 0;
 }
 
@@ -429,8 +431,8 @@ enum_scsi_adapters(void)
                 }
             } else {
                 err = GetLastError();
-                pr2serr("%s: IOCTL_SCSI_GET_INQUIRY_DATA failed err=%lu\n"
-                        "\t%s", adapter_name, err,
+                pr2serr("%s: IOCTL_SCSI_GET_INQUIRY_DATA failed err=%u\n"
+                        "\t%s", adapter_name, (unsigned int)err,
                         get_err_str(err, sizeof(b), b));
             }
             CloseHandle(fh);
@@ -440,8 +442,8 @@ enum_scsi_adapters(void)
                 fprintf(stderr, "%s: in use by other process (sharing "
                         "violation [34])\n", adapter_name);
             else if (verbose > 3)
-                pr2serr("%s: CreateFile failed err=%lu\n\t%s", adapter_name,
-                        err, get_err_str(err, sizeof(b), b));
+                pr2serr("%s: CreateFile failed err=%u\n\t%s", adapter_name,
+                        (unsigned int)err, get_err_str(err, sizeof(b), b));
             if (++hole_count >= MAX_HOLE_COUNT)
                 break;
         }
@@ -524,8 +526,8 @@ enum_pds(void)
                 fprintf(stderr, "%s: in use by other process (sharing "
                         "violation [34])\n", adapter_name);
             else if (verbose > 3)
-                pr2serr("%s: CreateFile failed err=%lu\n\t%s", adapter_name,
-                        err, get_err_str(err, sizeof(b), b));
+                pr2serr("%s: CreateFile failed err=%u\n\t%s", adapter_name,
+                        (unsigned int)err, get_err_str(err, sizeof(b), b));
             if (++hole_count >= MAX_HOLE_COUNT)
                 break;
         }
@@ -572,8 +574,8 @@ enum_cdroms(void)
                 fprintf(stderr, "%s: in use by other process (sharing "
                         "violation [34])\n", adapter_name);
             else if (verbose > 3)
-                pr2serr("%s: CreateFile failed err=%lu\n\t%s", adapter_name,
-                        err, get_err_str(err, sizeof(b), b));
+                pr2serr("%s: CreateFile failed err=%u\n\t%s", adapter_name,
+                        (unsigned int)err, get_err_str(err, sizeof(b), b));
             if (++hole_count >= MAX_HOLE_COUNT)
                 break;
         }
@@ -620,8 +622,8 @@ enum_tapes(void)
                 fprintf(stderr, "%s: in use by other process (sharing "
                         "violation [34])\n", adapter_name);
             else if (verbose > 3)
-                pr2serr("%s: CreateFile failed err=%lu\n\t%s", adapter_name,
-                        err, get_err_str(err, sizeof(b), b));
+                pr2serr("%s: CreateFile failed err=%u\n\t%s", adapter_name,
+                        (unsigned int)err, get_err_str(err, sizeof(b), b));
             if (++hole_count >= MAX_HOLE_COUNT)
                 break;
         }
